% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PersonAlyticsSim.R
\name{.active}
\alias{.active}
\title{.active - active bindings for Sim R6 class definitions}
\usage{
.active()
}
\description{
.active - active bindings for Sim R6 class definitions
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
\keyword{internal}
