% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{.assignNodes}
\alias{.assignNodes}
\title{.assignNodes - assign jobs to nodes}
\usage{
.assignNodes(condGrid, dirNames, nodeNames, seed)
}
\description{
.assignNodes - assign jobs to nodes
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
\keyword{internal}
