% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{.makeDirs}
\alias{.makeDirs}
\title{.makeDirs - create directories and return names}
\usage{
.makeDirs(condGrid, dirNames)
}
\description{
.makeDirs - create directories and return names
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
\keyword{internal}
