% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{.scanNode}
\alias{.scanNode}
\title{.scanNode - scan directories for this node and check how many are done}
\usage{
.scanNode(self, thisNode, theseJobs, ext = "complete")
}
\description{
.scanNode - scan directories for this node and check how many are done
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
\keyword{internal}
