% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PersonAlyticsSim.R
\docType{package}
\name{PersonAlyticsSim}
\alias{PersonAlyticsSim}
\title{PersonAlyticsSim: Tools for large scale simulations using PersonAlyticsPower}
\details{
PersonAlyticsSim has two basic functions: \code{\link{simDeploy}} and
\code{\link{simCollect}}. The first aids with setting up and running a
simulations, while the second is used to monitor progress and summarize and
plot the results.
}
\author{
Stephen Tueller \email{stueller@rti.org}
}
