% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PersonAlyticsSim.R
\docType{class}
\name{Sim}
\alias{Sim}
\title{\code{Sim} class generator}
\description{
\code{Sim} class generator

\code{Sim} class generator
}
\details{
Some text here
}
\section{Methods}{

\describe{
  \item{print}
}
}

\examples{

\dontrun{

}

}
\author{
Stephen Tueller \email{stueller@rti.org}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{simName}}{Charecter. A name for your simulation study.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Sim$new()}}
\item \href{#method-print}{\code{Sim$print()}}
\item \href{#method-emptyDirs}{\code{Sim$emptyDirs()}}
\item \href{#method-start}{\code{Sim$start()}}
\item \href{#method-cleandirs}{\code{Sim$cleandirs()}}
\item \href{#method-progress}{\code{Sim$progress()}}
\item \href{#method-summary}{\code{Sim$summary()}}
\item \href{#method-summarize}{\code{Sim$summarize()}}
\item \href{#method-glm}{\code{Sim$glm()}}
\item \href{#method-plot}{\code{Sim$plot()}}
\item \href{#method-clone}{\code{Sim$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sim$new(
  simName,
  seed,
  B,
  conditions = list(),
  dirNames = list(),
  type = c("param", "nonpar"),
  nodeNames = Sys.info()["nodename"]
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sim$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-emptyDirs"></a>}}
\if{latex}{\out{\hypertarget{method-emptyDirs}{}}}
\subsection{Method \code{emptyDirs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sim$emptyDirs()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-start"></a>}}
\if{latex}{\out{\hypertarget{method-start}{}}}
\subsection{Method \code{start()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sim$start(
  ext = "complete",
  run1 = NULL,
  cleanup = TRUE,
  checkOnly = FALSE,
  debug = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{checkOnly}}{Not implemented, consider deprecating, $progress() works}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cleandirs"></a>}}
\if{latex}{\out{\hypertarget{method-cleandirs}{}}}
\subsection{Method \code{cleandirs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sim$cleandirs()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-progress"></a>}}
\if{latex}{\out{\hypertarget{method-progress}{}}}
\subsection{Method \code{progress()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sim$progress(ext = "complete", ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-summary"></a>}}
\if{latex}{\out{\hypertarget{method-summary}{}}}
\subsection{Method \code{summary()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sim$summary(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-summarize"></a>}}
\if{latex}{\out{\hypertarget{method-summarize}{}}}
\subsection{Method \code{summarize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sim$summarize(fpc = TRUE, raw = FALSE, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-glm"></a>}}
\if{latex}{\out{\hypertarget{method-glm}{}}}
\subsection{Method \code{glm()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sim$glm(
  y,
  interactionOrder = 1,
  alpha = 0.05,
  subset = NULL,
  subsetVarName = NULL,
  recode = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\if{latex}{\out{\hypertarget{method-plot}{}}}
\subsection{Method \code{plot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sim$plot(
  x,
  y,
  group,
  facet = NULL,
  subset = NULL,
  title = "",
  data = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sim$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
